Type audio language detector metadata
